Cleanup of code after simulatorservice splitting
