+++ src/main/java/com/zuehlke/carrera/simulator/communication/announce/AutonomousSimulatorAnnouncer.java
@@ -1,6 +1,6 @@
 package com.zuehlke.carrera.simulator.communication.announce;
 
-import com.zuehlke.carrera.simulator.domain.api.SimulatorApiAdapter;
+import com.zuehlke.carrera.simulator.services.adapter.SimulatorApiAdapter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Component;
 @EnableScheduling
 public class AutonomousSimulatorAnnouncer {
     private final SimulatorApiAdapter apiAdapter;
-    private final WhereAmIHostedService whereAmIHostedService;
+    private final WhereAmIHosted whereAmIHostedService;
 
     @Autowired
-    public AutonomousSimulatorAnnouncer(SimulatorApiAdapter apiAdapter, WhereAmIHostedService whereAmIHosted) {
+    public AutonomousSimulatorAnnouncer(SimulatorApiAdapter apiAdapter, WhereAmIHosted whereAmIHosted) {
         this.apiAdapter = apiAdapter;
         this.whereAmIHostedService = whereAmIHosted;
     }
