+++ src/main/java/com/zuehlke/carrera/simulator/config/RabbitCommunicationProtocolConfig.java
@@ -8,9 +8,9 @@ import com.zuehlke.carrera.api.client.Client;
 import com.zuehlke.carrera.api.client.rabbit.RabbitClient;
 import com.zuehlke.carrera.api.seralize.JacksonSerializer;
 import com.zuehlke.carrera.api.seralize.Serializer;
-import com.zuehlke.carrera.simulator.domain.api.RabbitSimulatorApiAdapter;
-import com.zuehlke.carrera.simulator.domain.api.SimulatorApiAdapter;
-import com.zuehlke.carrera.simulator.domain.simulib.SimulibRabbitApiAdapter;
+import com.zuehlke.carrera.simulator.services.adapter.RabbitSimulatorApiAdapter;
+import com.zuehlke.carrera.simulator.services.adapter.SimulatorApiAdapter;
+import com.zuehlke.carrera.simulator.services.adapter.RabbitSimulibApiAdapter;
 import com.zuehlke.carrera.simulator.model.PilotInterface;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -27,7 +27,7 @@ public class RabbitCommunicationProtocolConfig {
 
     @Bean
     public PilotInterface pilotInterface(SimulatorApi simulatorApi) {
-        return new SimulibRabbitApiAdapter(simulatorApi);
+        return new RabbitSimulibApiAdapter(simulatorApi);
     }
 
     @Bean
