+++ src/main/java/com/zuehlke/carrera/simulator/web/SimulatorRestController.java
@@ -1,7 +1,7 @@
 package com.zuehlke.carrera.simulator.web;
 
 import com.zuehlke.carrera.simulator.model.racetrack.TrackInfo;
-import com.zuehlke.carrera.simulator.domain.SimulatorService;
+import com.zuehlke.carrera.simulator.services.SimulatorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,11 +14,6 @@ public class SimulatorRestController {
     @Autowired
     public SimulatorRestController(SimulatorService simulatorService) {
         this.simulatorService = simulatorService;
-    }
-
-    @RequestMapping(value = "/track", method = RequestMethod.GET, produces = "application/json")
-    public TrackInfo getTrack() {
-        return simulatorService.getTrackInfo();
     }
 
     @RequestMapping(value = "/start", method = RequestMethod.POST)
@@ -37,13 +32,18 @@ public class SimulatorRestController {
     }
 
     @RequestMapping(value = "/powerup/{delta}", method = RequestMethod.POST)
-    public void powerup(@PathVariable int delta) {
-        simulatorService.powerup(delta);
+    public void powerUp(@PathVariable int delta) {
+        simulatorService.powerUp(delta);
     }
 
     @RequestMapping(value = "/powerdown/{delta}", method = RequestMethod.POST)
-    public void powerdown(@PathVariable int delta) {
-        simulatorService.powerdown(delta);
+    public void powerDown(@PathVariable int delta) {
+        simulatorService.powerDown(delta);
+    }
+
+    @RequestMapping(value = "/track", method = RequestMethod.GET, produces = "application/json")
+    public TrackInfo getTrack() {
+        return simulatorService.getTrackInfo();
     }
 
     @RequestMapping(value = "/selectDesign", method = RequestMethod.POST, produces = "application/json")
